requirements.txt
binance.py
python-binance
binance-api
python-dotenv
pandas
